requirements/requirement.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.2
django-ckeditor==6.5.1
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
filelock==3.12.0
gunicorn==20.1.0
Markdown==3.4.3
platformdirs==3.5.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
